requirements.txt
google-generativeai==0.3.2
flask==3.0.2
python-dotenv==1.0.1
Pillow==10.2.0
flask-socketio==5.3.6
eventlet==0.35.2
Flask-Session==0.6.0 